agno/requirements.txt
# Agno AI Agent Framework
# Docs: https://docs.agno.com
agno[os]>=1.4.0

# LLM Providers
anthropic>=0.49.0
openai>=1.75.0
google-generativeai>=0.8.0

# Tools & Integrations
mcp>=1.6.0

# Database (session persistence)
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.39

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Utils
python-dotenv>=1.0.0
httpx>=0.28.0
pydantic>=2.11.0
